Skeleton of progress bars

+++ DupFinderCore/ImagesLoadedProgress.cs
@@ -4,6 +4,6 @@
     {
         public int TotalImages { get; set; }
         public int AmountDone { get; set; }
-        public int PercentageDone => AmountDone * 100 / TotalImages;
+        public int PercentageDone => (int)(AmountDone * 100 / (double)TotalImages);
     }
 }
